requirements.txt
asgiref==3.11.0
Django==5.0
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
pillow==12.0.0
psycopg2-binary==2.9.11
PyJWT==2.10.1
python-decouple==3.8
setuptools==80.9.0
sqlparse==0.5.4
tzdata==2025.3
wheel==0.45.1
